requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
flatbuffers==1.12
gast==0.3.3
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.4
importlib-metadata==6.0.0
Keras-Preprocessing==1.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
Pillow==9.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.11.0
